Sampled Stream Function
With this new version there is the possibility of streaming tweets from the sampled stream endpoint from Twitter. This is experimental and might still contain some errors. Stream gets stored as a JSON file.

Additionally this update provides a parsing function for the streamed json file. This function ist slow and will most likely fail with long streams.